+++ server/controllers/appointments/appointmentDetailsController.test.ts
@@ -4,8 +4,7 @@ import CheckAppointmentDetailsPage from '../../pages/appointments/checkAppointme
 import AppointmentService from '../../services/appointmentService'
 import appointmentFactory from '../../testutils/factories/appointmentFactory'
 import ProjectDetailsController from './appointmentDetailsController'
-import AppointmentFormService from '../../services/forms/appointmentFormService'
-import { AppointmentOutcomeForm } from '../../@types/user-defined'
+import AppointmentFormService, { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import appointmentOutcomeFormFactory from '../../testutils/factories/appointmentOutcomeFormFactory'
 import ProjectService from '../../services/projectService'
 import projectFactory from '../../testutils/factories/projectFactory'
