+++ server/controllers/appointments/attendanceOutcomeController.ts
@@ -5,13 +5,12 @@ import AttendanceOutcomePage, {
   AttendanceOutcomeBody,
   AttendanceOutcomeContext,
 } from '../../pages/appointments/attendanceOutcomePage'
-import AppointmentFormService from '../../services/forms/appointmentFormService'
+import AppointmentFormService, { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import SessionService from '../../services/sessionService'
 import BaseAppointmentController, {
   AppointmentStepViewDataParams,
   ContextDataParams,
 } from './baseAppointmentController'
-import { AppointmentOutcomeForm } from '../../@types/user-defined'
 import { AppointmentDto } from '../../@types/shared'
 
 export default class AttendanceOutcomeController extends BaseAppointmentController<AttendanceOutcomePage> {
