+++ server/controllers/appointments/baseAppointmentController.ts
@@ -1,12 +1,11 @@
 import type { Request, RequestHandler, Response } from 'express'
 import BaseAppointmentUpdatePage from '../../pages/appointments/baseAppointmentUpdatePage'
 import AppointmentService from '../../services/appointmentService'
-import AppointmentFormService from '../../services/forms/appointmentFormService'
+import AppointmentFormService, { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import SessionService from '../../services/sessionService'
 import {
   AppointmentOrSessionParams,
   AppointmentOrSession,
-  AppointmentOutcomeForm,
   ValidationErrors,
   IFormPageController,
 } from '../../@types/user-defined'
