+++ server/controllers/appointments/bulkUpdateController.ts
@@ -1,11 +1,11 @@
 import type { Request, RequestHandler, Response } from 'express'
 
-import AppointmentFormService from '../../services/forms/appointmentFormService'
-import { AppointmentOutcomeForm, IFormPageController } from '../../@types/user-defined'
+import AppointmentFormService, { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import BulkUpdatePage from '../../pages/appointments/bulkUpdatePage'
 import SessionService from '../../services/sessionService'
 import AppointmentService from '../../services/appointmentService'
 import ProjectService from '../../services/projectService'
+import { IFormPageController } from '../../@types/user-defined'
 
 export default class BulkUpdateController implements IFormPageController {
   constructor(
