+++ server/pages/appointments/confirmPage.ts
@@ -2,12 +2,12 @@ import { AppointmentDto, ContactOutcomeDto } from '../../@types/shared'
 import {
   AppointmentOrSession,
   AppointmentOrSessionParams,
-  AppointmentOutcomeForm,
   GovUkRadioOrCheckboxOption,
   GovUkSummaryListItem,
   ValidationErrors,
   YesOrNo,
 } from '../../@types/user-defined'
+import { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import GovUkRadioGroup from '../../forms/GovUkRadioGroup'
 import Offender from '../../models/offender'
 import AppointmentUtils from '../../utils/appointmentUtils'
