+++ server/pages/appointments/logHoursPage.ts
@@ -1,9 +1,5 @@
-import {
-  AppointmentOrSessionParams,
-  AppointmentOutcomeForm,
-  AppointmentUpdateQuery,
-  ValidationErrors,
-} from '../../@types/user-defined'
+import { AppointmentOrSessionParams, AppointmentUpdateQuery, ValidationErrors } from '../../@types/user-defined'
+import { AppointmentOutcomeForm } from '../../services/forms/appointmentFormService'
 import DateTimeFormats from '../../utils/dateTimeUtils'
 import BaseAppointmentUpdatePage from './baseAppointmentUpdatePage'
 import { AppointmentFormPage } from './pathMap'
@@ -65,7 +61,7 @@ export default class LogHoursPage extends BaseAppointmentUpdatePage<LogHoursBody
     }
   }
 
-  protected backPage(_params: AppointmentOrSessionParams): AppointmentFormPage {
+  protected backPage(_appointmentOrSession: AppointmentOrSessionParams): AppointmentFormPage {
     return 'attendance-outcome'
   }
